apps/backend/requirements.txt
# FastAPI Backend Dependencies

# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
email-validator>=2.0.0

# Database - MongoDB
motor>=3.3.0  # Async MongoDB driver
pymongo>=4.6.0
beanie>=1.24.0  # ODM for MongoDB

# Vector Database
qdrant-client>=1.7.0  # Vector search
sentence-transformers>=2.2.2  # Embeddings

# Async support
asyncio>=3.4.3
aiofiles>=23.2.1

# Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# AI & LLM
openai>=1.12.0
google-generativeai>=0.3.2  # Gemini
groq>=0.4.1
langchain>=0.1.0
langchain-community>=0.0.20

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
markdown>=3.5.0

# Web Scraping & Automation
beautifulsoup4>=4.12.0
selenium>=4.16.0
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.3

# Monitoring (optional)
prometheus-client>=0.20.0

# Development
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0  # For testing
